feat: enhance object management with new Objet interface and improve API request handling
